requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
pandas>=2.0.0
requests>=2.31.0

# ── Google Drive ──────────────────────────────────────────────────────────────
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1

# ── OCR ───────────────────────────────────────────────────────────────────────
Pillow>=10.0.0
pdfminer.six>=20221105
# pytesseract>=0.3.10  # optionnel, si Tesseract installé localement
